[Refactor] #120 - CalendarViewModel 생성 & Extension 분리
